# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Task API",
        "version": "1.0",
        "description": "In-memory CRUD API for managing tasks."
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "summary": "API info",
                "operationId": "getRoot",
                "responses": {
                    "200": {
                        "description": "Basic API metadata",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "name",
                                        "version",
                                        "endpoints"
                                    ],
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Task API"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "1.0"
                                        },
                                        "endpoints": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "/tasks"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Health check",
                "operationId": "getHealth",
                "responses": {
                    "200": {
                        "description": "Service is up",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "status"
                                    ],
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "summary": "List tasks",
                "operationId": "listTasks",
                "responses": {
                    "200": {
                        "description": "All tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a task",
                "operationId": "createTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "examples": {
                                    "wrongType": {
                                        "value": {
                                            "error": "title must be a string"
                                        }
                                    },
                                    "emptyTitle": {
                                        "value": {
                                            "error": "title is required and cannot be empty"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "Task id",
                    "schema": {
                        "type": "integer",
                        "example": 1
                    }
                }
            ],
            "get": {
                "summary": "Get one task",
                "operationId": "getTask",
                "responses": {
                    "200": {
                        "description": "Task found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": "Task 99 not found"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update a task",
                "operationId": "updateTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "examples": {
                                    "emptyBody": {
                                        "value": {
                                            "error": "request body must include title and/or done"
                                        }
                                    },
                                    "invalidTitle": {
                                        "value": {
                                            "error": "title must be a non-empty string"
                                        }
                                    },
                                    "invalidDone": {
                                        "value": {
                                            "error": "done must be a boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": "Task 99 not found"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete a task",
                "operationId": "deleteTask",
                "responses": {
                    "204": {
                        "description": "Task deleted"
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "example": {
                                    "error": "Task 99 not found"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "done"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Learn Express basics"
                    },
                    "done": {
                        "type": "boolean",
                        "example": true
                    }
                }
            },
            "CreateTaskRequest": {
                "type": "object",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Build the Task API"
                    }
                }
            },
            "UpdateTaskRequest": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Read a book"
                    },
                    "done": {
                        "type": "boolean",
                        "example": true
                    }
                }
            },
            "Error": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            }
        }
    }
}